move cluster session to mars.session
